@@ -1,6 +1,3 @@
-use std::env;
-use std::ffi::OsString;
-use std::io::ErrorKind;
 use std::process;
 use std::process::Stdio;
 use std::str::FromStr;
@@ -8,10 +5,10 @@ use std::str::FromStr;
 use crate::config::ModuleConfig;
 use crate::config::StarshipConfig;
 use crate::configs::PROMPT_ORDER;
+use crate::context::Context;
 use crate::utils;
 use std::fs::File;
 use std::io::Write;
-use toml::Value;
 use toml_edit::Document;
 
 #[cfg(not(windows))]
@@ -19,15 +16,15 @@ const STD_EDITOR: &str = "vi";
 #[cfg(windows)]
 const STD_EDITOR: &str = "notepad.exe";
 
-pub fn update_configuration(name: &str, value: &str) {
-    let mut doc = get_configuration_edit();
+pub fn update_configuration(context: &Context, name: &str, value: &str) {
+    let mut doc = get_configuration_edit(context);
 
     match handle_update_configuration(&mut doc, name, value) {
         Err(e) => {
             eprintln!("{e}");
             process::exit(1);
         }
-        _ => write_configuration(&doc),
+        _ => write_configuration(context, &doc),
     }
 }
 
@@ -72,7 +69,7 @@ fn handle_update_configuration(doc: &mut Document, name: &str, value: &str) -> R
     Ok(())
 }
 
-pub fn print_configuration(use_default: bool, paths: &[String]) {
+pub fn print_configuration(context: &Context, use_default: bool, paths: &[String]) -> String {
     let config = if use_default {
         // Get default config
         let default_config = crate::configs::FullConfig::default();
@@ -80,7 +77,7 @@ pub fn print_configuration(use_default: bool, paths: &[String]) {
         toml::value::Value::try_from(default_config).unwrap()
     } else {
         // Get config as toml::Value
-        let user_config = get_configuration();
+        let user_config = get_configuration(context);
         // Convert into FullConfig and fill in default values
         let user_config = crate::configs::FullConfig::load(&user_config);
         // Convert back to Value because toml can't serialize FullConfig directly
@@ -125,14 +122,13 @@ pub fn print_configuration(use_default: bool, paths: &[String]) {
     let string_config = toml::to_string_pretty(&print_config).unwrap();
 
     println!("{string_config}");
+    string_config
 }
 
 fn extract_toml_paths(mut config: toml::Value, paths: &[String]) -> toml::Value {
     // Extract all the requested sections into a new configuration.
     let mut subset = toml::value::Table::new();
-    let config = if let Some(config) = config.as_table_mut() {
-        config
-    } else {
+    let Some(config) = config.as_table_mut() else {
         // This function doesn't make any sense if the root is not a table.
         return toml::Value::Table(subset);
     };
@@ -156,9 +152,7 @@ fn extract_toml_paths(mut config: toml::Value, paths: &[String]) -> toml::Value 
         }
 
         // Extract the value to move.
-        let value = if let Some(value) = source_cursor.remove(end) {
-            value
-        } else {
+        let Some(value) = source_cursor.remove(end) else {
             // We didn't find a value for this path, so move on to the next path.
             continue 'paths;
         };
@@ -181,15 +175,15 @@ fn extract_toml_paths(mut config: toml::Value, paths: &[String]) -> toml::Value 
     toml::Value::Table(subset)
 }
 
-pub fn toggle_configuration(name: &str, key: &str) {
-    let mut doc = get_configuration_edit();
+pub fn toggle_configuration(context: &Context, name: &str, key: &str) {
+    let mut doc = get_configuration_edit(context);
 
     match handle_toggle_configuration(&mut doc, name, key) {
         Err(e) => {
             eprintln!("{e}");
             process::exit(1);
         }
-        _ => write_configuration(&doc),
+        _ => write_configuration(context, &doc),
     }
 }
 
@@ -222,32 +216,15 @@ fn handle_toggle_configuration(doc: &mut Document, name: &str, key: &str) -> Res
     Ok(())
 }
 
-pub fn get_configuration() -> Value {
-    let starship_config = StarshipConfig::initialize();
+pub fn get_configuration(context: &Context) -> toml::Table {
+    let starship_config = StarshipConfig::initialize(&context.get_config_path_os());
 
-    starship_config
-        .config
-        .expect("Failed to load starship config")
+    starship_config.config.unwrap_or_default()
 }
 
-pub fn get_configuration_edit() -> Document {
-    let file_path = get_config_path();
-    let toml_content = match utils::read_file(file_path) {
-        Ok(content) => {
-            log::trace!("Config file content: \"\n{}\"", &content);
-            Some(content)
-        }
-        Err(e) => {
-            let level = if e.kind() == ErrorKind::NotFound {
-                log::Level::Debug
-            } else {
-                log::Level::Error
-            };
-
-            log::log!(level, "Unable to read config file content: {}", &e);
-            None
-        }
-    };
+pub fn get_configuration_edit(context: &Context) -> Document {
+    let config_file_path = context.get_config_path_os();
+    let toml_content = StarshipConfig::read_config_content_as_str(&config_file_path);
 
     toml_content
         .unwrap_or_default()
@@ -255,8 +232,11 @@ pub fn get_configuration_edit() -> Document {
         .expect("Failed to load starship config")
 }
 
-pub fn write_configuration(doc: &Document) {
-    let config_path = get_config_path();
+pub fn write_configuration(context: &Context, doc: &Document) {
+    let config_path = context.get_config_path_os().unwrap_or_else(|| {
+        eprintln!("config path required to write configuration");
+        process::exit(1);
+    });
 
     let config_str = doc.to_string();
 
@@ -265,10 +245,16 @@ pub fn write_configuration(doc: &Document) {
         .expect("Error writing starship config");
 }
 
-pub fn edit_configuration(editor_override: Option<&str>) -> Result<(), Box<dyn std::error::Error>> {
+pub fn edit_configuration(
+    context: &Context,
+    editor_override: Option<&str>,
+) -> Result<(), Box<dyn std::error::Error>> {
     // Argument currently only used for testing, but could be used to specify
     // an editor override on the command line.
-    let config_path = get_config_path();
+    let config_path = context.get_config_path_os().unwrap_or_else(|| {
+        eprintln!("config path required to edit configuration");
+        process::exit(1);
+    });
 
     let editor_cmd = shell_words::split(&get_editor(editor_override))?;
     let mut command = match utils::create_command(&editor_cmd[0]) {
@@ -318,19 +304,18 @@ fn get_editor_internal(visual: Option<String>, editor: Option<String>) -> String
     STD_EDITOR.into()
 }
 
-fn get_config_path() -> OsString {
-    if let Some(config_path) = env::var_os("STARSHIP_CONFIG") {
-        return config_path;
-    }
-    utils::home_dir()
-        .expect("couldn't find home directory")
-        .join(".config")
-        .join("starship.toml")
-        .into()
-}
-
 #[cfg(test)]
 mod tests {
+    use std::{fs::create_dir, io};
+
+    use tempfile::TempDir;
+    use toml_edit::Item;
+
+    use crate::{
+        context::{Shell, Target},
+        context_env::Env,
+    };
+
     use super::*;
 
     // This is every possible permutation, 3² = 9.
@@ -383,14 +368,14 @@ mod tests {
     }
 
     #[test]
-    fn no_panic_when_editor_unparseable() {
-        let outcome = edit_configuration(Some("\"vim"));
+    fn no_panic_when_editor_unparsable() {
+        let outcome = edit_configuration(&Default::default(), Some("\"vim"));
         assert!(outcome.is_err());
     }
 
     #[test]
     fn no_panic_when_editor_not_found() {
-        let outcome = edit_configuration(Some("this_editor_does_not_exist"));
+        let outcome = edit_configuration(&Default::default(), Some("this_editor_does_not_exist"));
         assert!(outcome.is_err());
     }
 
@@ -431,7 +416,7 @@ mod tests {
             ok = true
         };
         let actual_config = extract_toml_paths(
-            config,
+            toml::Value::Table(config),
             &[
                 "extract_root".to_owned(),
                 "extract_section".to_owned(),
@@ -439,7 +424,7 @@ mod tests {
             ],
         );
 
-        assert_eq!(expected_config, actual_config);
+        assert_eq!(toml::Value::Table(expected_config), actual_config);
     }
 
     fn create_doc() -> Document {
@@ -585,5 +570,116 @@ mod tests {
         assert!(doc["a"]["b"]["c"]["d"]["e"]["f"]["g"]["h"]
             .as_bool()
             .unwrap())
+    }
+
+    #[test]
+    fn write_and_get_configuration_test() -> io::Result<()> {
+        let dir = tempfile::tempdir()?;
+        let context = setup_config(&dir, true, StarshipConfigEnvScenario::NotSpecified)?;
+        let mut doc = get_configuration_edit(&context);
+        doc["directory"]["format"] = Item::Value("myformat".into());
+        write_configuration(&context, &doc);
+        let doc_reloaded = get_configuration_edit(&context);
+        assert_eq!(
+            "myformat",
+            doc_reloaded["directory"]["format"].as_str().unwrap()
+        );
+        dir.close()
+    }
+
+    const PRINT_CONFIG_DEFAULT: &str = "[custom]";
+    const PRINT_CONFIG_HOME: &str = "[custom.home]";
+    const PRINT_CONFIG_ENV: &str = "[custom.STARSHIP_CONFIG]";
+
+    #[test]
+    fn print_configuration_scenarios() -> io::Result<()> {
+        run_print_configuration_test(
+            "~/.config/starship.toml, no STARSHIP_CONFIG uses home",
+            true,
+            StarshipConfigEnvScenario::NotSpecified,
+            PRINT_CONFIG_HOME,
+        )?;
+        run_print_configuration_test(
+            "no ~/.config/starship.toml, no STARSHIP_CONFIG uses default",
+            false,
+            StarshipConfigEnvScenario::NotSpecified,
+            PRINT_CONFIG_DEFAULT,
+        )?;
+        run_print_configuration_test(
+            "~/.config/starship.toml, STARSHIP_CONFIG nonexisting file uses default",
+            true,
+            StarshipConfigEnvScenario::NonExistingFile,
+            PRINT_CONFIG_DEFAULT,
+        )?;
+        run_print_configuration_test(
+            "~/.config/starship.toml, STARSHIP_CONFIG existing file uses STARSHIP_CONFIG file",
+            true,
+            StarshipConfigEnvScenario::ExistingFile,
+            PRINT_CONFIG_ENV,
+        )?;
+        Ok(())
+    }
+
+    enum StarshipConfigEnvScenario {
+        NotSpecified,
+        NonExistingFile,
+        ExistingFile,
+    }
+
+    fn run_print_configuration_test(
+        message: &str,
+        home_file_exists: bool,
+        starship_config_env_scenario: StarshipConfigEnvScenario,
+        expected_first_line: &str,
+    ) -> io::Result<()> {
+        let dir = tempfile::tempdir()?;
+        let context = setup_config(&dir, home_file_exists, starship_config_env_scenario)?;
+        let config = print_configuration(&context, false, &["custom".to_string()]);
+        let first_line = config.split('\n').next().unwrap();
+        assert_eq!(expected_first_line, first_line, "{message}");
+        dir.close()
+    }
+
+    fn setup_config(
+        dir: &TempDir,
+        home_file_exists: bool,
+        starship_config_env_scenario: StarshipConfigEnvScenario,
+    ) -> io::Result<Context> {
+        let config_path = dir.path().to_path_buf().join(".config");
+        create_dir(&config_path)?;
+        let home_starship_toml = config_path.join("starship.toml");
+        let env_toml = dir.path().join("env.toml");
+        if home_file_exists {
+            let mut home_file = File::create(home_starship_toml)?;
+            home_file.write_all(PRINT_CONFIG_HOME.as_bytes())?;
+        }
+
+        let env_starship_config = match starship_config_env_scenario {
+            StarshipConfigEnvScenario::NotSpecified => None,
+            StarshipConfigEnvScenario::NonExistingFile => Some(env_toml),
+            StarshipConfigEnvScenario::ExistingFile => {
+                let mut env_toml_file = File::create(&env_toml)?;
+                env_toml_file.write_all(PRINT_CONFIG_ENV.as_bytes())?;
+                Some(env_toml)
+            }
+        };
+
+        let mut env = Env::default();
+        if let Some(v) = env_starship_config {
+            env.insert("STARSHIP_CONFIG", v.to_string_lossy().to_string());
+        }
+        env.insert(
+            "HOME",
+            dir.path().to_path_buf().to_string_lossy().to_string(),
+        );
+
+        Ok(Context::new_with_shell_and_path(
+            Default::default(),
+            Shell::Unknown,
+            Target::Main,
+            Default::default(),
+            Default::default(),
+            env,
+        ))
     }
 }
