@@ -1,6 +1,7 @@
 use crate::configs::Palette;
 use crate::context::Context;
-use crate::serde_utils::ValueDeserializer;
+
+use crate::serde_utils::{ValueDeserializer, ValueRef};
 use crate::utils;
 use nu_ansi_term::Color;
 use serde::{
@@ -10,9 +11,9 @@ use serde::{
 use std::borrow::Cow;
 use std::clone::Clone;
 use std::collections::HashMap;
+use std::ffi::OsString;
 use std::io::ErrorKind;
 
-use std::env;
 use toml::Value;
 
 /// Root config of a module.
@@ -22,12 +23,12 @@ where
     E: SerdeError,
 {
     /// Construct a `ModuleConfig` from a toml value.
-    fn from_config(config: &'a Value) -> Result<Self, E>;
+    fn from_config<V: Into<ValueRef<'a>>>(config: V) -> Result<Self, E>;
 
     /// Loads the TOML value into the config.
     /// Missing values are set to their default values.
     /// On error, logs an error message.
-    fn load(config: &'a Value) -> Self {
+    fn load<V: Into<ValueRef<'a>>>(config: V) -> Self {
         match Self::from_config(config) {
             Ok(config) => config,
             Err(e) => {
@@ -39,14 +40,15 @@ where
 
     /// Helper function that will call `ModuleConfig::from_config(config)  if config is Some,
     /// or `ModuleConfig::default()` if config is None.
-    fn try_load(config: Option<&'a Value>) -> Self {
-        config.map(Self::load).unwrap_or_default()
+    fn try_load<V: Into<ValueRef<'a>>>(config: Option<V>) -> Self {
+        config.map(Into::into).map(Self::load).unwrap_or_default()
     }
 }
 
 impl<'a, T: Deserialize<'a> + Default> ModuleConfig<'a, ValueError> for T {
     /// Create `ValueDeserializer` wrapper and use it to call `Deserialize::deserialize` on it.
-    fn from_config(config: &'a Value) -> Result<Self, ValueError> {
+    fn from_config<V: Into<ValueRef<'a>>>(config: V) -> Result<Self, ValueError> {
+        let config = config.into();
         let deserializer = ValueDeserializer::new(config);
         T::deserialize(deserializer).or_else(|err| {
             // If the error is an unrecognized key, print a warning and run
@@ -114,44 +116,46 @@ where
 }
 
 /// Root config of starship.
+#[derive(Default)]
 pub struct StarshipConfig {
-    pub config: Option<Value>,
-}
-
-pub fn get_config_path() -> Option<String> {
-    if let Ok(path) = env::var("STARSHIP_CONFIG") {
-        // Use $STARSHIP_CONFIG as the config path if available
-        log::debug!("STARSHIP_CONFIG is set: {}", &path);
-        Some(path)
-    } else {
-        // Default to using ~/.config/starship.toml
-        log::debug!("STARSHIP_CONFIG is not set");
-        let config_path = utils::home_dir()?.join(".config/starship.toml");
-        let config_path_str = config_path.to_str()?.to_owned();
-        log::debug!("Using default config path: {}", config_path_str);
-        Some(config_path_str)
-    }
+    pub config: Option<toml::Table>,
 }
 
 impl StarshipConfig {
     /// Initialize the Config struct
-    pub fn initialize() -> Self {
-        if let Some(file_data) = Self::config_from_file() {
-            Self {
-                config: Some(file_data),
+    pub fn initialize(config_file_path: &Option<OsString>) -> Self {
+        Self::config_from_file(config_file_path)
+            .map(|config| Self {
+                config: Some(config),
+            })
+            .unwrap_or_default()
+    }
+
+    /// Create a config from a starship configuration file
+    fn config_from_file(config_file_path: &Option<OsString>) -> Option<toml::Table> {
+        let toml_content = Self::read_config_content_as_str(config_file_path)?;
+
+        match toml::from_str(&toml_content) {
+            Ok(parsed) => {
+                log::debug!("Config parsed: {:?}", &parsed);
+                Some(parsed)
             }
-        } else {
-            Self {
-                config: Some(Value::Table(toml::value::Table::new())),
+            Err(error) => {
+                log::error!("Unable to parse the config file: {}", error);
+                None
             }
         }
     }
 
-    /// Create a config from a starship configuration file
-    fn config_from_file() -> Option<Value> {
-        let file_path = get_config_path()?;
-
-        let toml_content = match utils::read_file(file_path) {
+    pub fn read_config_content_as_str(config_file_path: &Option<OsString>) -> Option<String> {
+        if config_file_path.is_none() {
+            log::debug!(
+                "Unable to determine `config_file_path`. Perhaps `utils::home_dir` is not defined on your platform?"
+            );
+            return None;
+        }
+        let config_file_path = config_file_path.as_ref().unwrap();
+        match utils::read_file(config_file_path) {
             Ok(content) => {
                 log::trace!("Config file content: \"\n{}\"", &content);
                 Some(content)
@@ -164,17 +168,6 @@ impl StarshipConfig {
                 };
 
                 log::log!(level, "Unable to read config file content: {}", &e);
-                None
-            }
-        }?;
-
-        match toml::from_str(&toml_content) {
-            Ok(parsed) => {
-                log::debug!("Config parsed: {:?}", &parsed);
-                Some(parsed)
-            }
-            Err(error) => {
-                log::error!("Unable to parse the config file: {}", error);
                 None
             }
         }
@@ -195,7 +188,7 @@ impl StarshipConfig {
 
     /// Get the value of the config in a specific path
     pub fn get_config(&self, path: &[&str]) -> Option<&Value> {
-        let mut prev_table = self.config.as_ref()?.as_table()?;
+        let mut prev_table = self.config.as_ref()?;
 
         assert_ne!(
             path.len(),
@@ -922,5 +915,14 @@ mod tests {
 
         // Test default behavior
         assert!(get_palette(&palettes, None).is_none());
+    }
+
+    #[test]
+    fn read_config_no_config_file_path_provided() {
+        assert_eq!(
+            None,
+            StarshipConfig::read_config_content_as_str(&None),
+            "if the platform doesn't have utils::home_dir(), it should return None"
+        );
     }
 }
