@@ -22,11 +22,11 @@ pub struct BatteryConfig<'a> {
 impl<'a> Default for BatteryConfig<'a> {
     fn default() -> Self {
         BatteryConfig {
-            full_symbol: " ",
-            charging_symbol: " ",
-            discharging_symbol: " ",
-            unknown_symbol: " ",
-            empty_symbol: " ",
+            full_symbol: "󰁹 ",
+            charging_symbol: "󰂄 ",
+            discharging_symbol: "󰂃 ",
+            unknown_symbol: "󰁽 ",
+            empty_symbol: "󰂎 ",
             format: "[$symbol$percentage]($style) ",
             display: vec![BatteryDisplayConfig::default()],
             disabled: false,
