@@ -18,7 +18,7 @@ pub struct OSConfig<'a> {
 
 impl<'a> OSConfig<'a> {
     pub fn get_symbol(&self, key: &Type) -> Option<&'a str> {
-        self.symbols.get(key).cloned()
+        self.symbols.get(key).copied()
     }
 }
 
@@ -28,10 +28,12 @@ impl<'a> Default for OSConfig<'a> {
             format: "[$symbol]($style)",
             style: "bold white",
             symbols: indexmap! {
+                Type::Alpaquita => "🔔 ",
                 Type::Alpine => "🏔️ ",
                 Type::Amazon => "🙂 ",
                 Type::Android => "🤖 ",
                 Type::Arch => "🎗️ ",
+                Type::Artix => "🎗️ ",
                 Type::CentOS => "💠 ",
                 Type::Debian => "🌀 ",
                 Type::DragonFly => "🐉 ",
@@ -44,6 +46,7 @@ impl<'a> Default for OSConfig<'a> {
                 Type::HardenedBSD => "🛡️ ",
                 Type::Illumos => "🐦 ",
                 Type::Linux => "🐧 ",
+                Type::Mabox => "📦 ",
                 Type::Macos => "🍎 ",
                 Type::Manjaro => "🥭 ",
                 Type::Mariner => "🌊 ",
@@ -52,6 +55,8 @@ impl<'a> Default for OSConfig<'a> {
                 Type::NetBSD => "🚩 ",
                 Type::NixOS => "❄️ ",
                 Type::OpenBSD => "🐡 ",
+                Type::OpenCloudOS => "☁️ ",
+                Type::openEuler => "🦉 ",
                 Type::openSUSE => "🦎 ",
                 Type::OracleLinux => "🦴 ",
                 Type::Pop => "🍭 ",
