@@ -1,20 +1,17 @@
-use unicode_segmentation::UnicodeSegmentation;
+use std::io::Error;
+use std::path::Path;
 
+use super::utils::truncate::truncate_text;
 use super::{Context, Module, ModuleConfig};
 
 use crate::configs::hg_branch::HgBranchConfig;
 use crate::formatter::StringFormatter;
+use crate::utils::read_file;
 
 /// Creates a module with the Hg bookmark or branch in the current directory
 ///
 /// Will display the bookmark or branch name if the current directory is an hg repo
 pub fn module<'a>(context: &'a Context) -> Option<Module<'a>> {
-    let is_hg_repo = context.try_begin_scan()?.set_folders(&[".hg"]).is_match();
-
-    if !is_hg_repo {
-        return None;
-    }
-
     let mut module = context.new_module("hg_branch");
     let config: HgBranchConfig = HgBranchConfig::try_load(module.config);
 
@@ -34,16 +31,16 @@ pub fn module<'a>(context: &'a Context) -> Option<Module<'a>> {
         config.truncation_length as usize
     };
 
-    let branch_name =
-        get_hg_current_bookmark(context).unwrap_or_else(|| get_hg_branch_name(context));
+    let repo_root = context.begin_ancestor_scan().set_folders(&[".hg"]).scan()?;
+    let branch_name = get_hg_current_bookmark(repo_root).unwrap_or_else(|_| {
+        get_hg_branch_name(repo_root).unwrap_or_else(|_| String::from("default"))
+    });
 
-    let truncated_graphemes = get_graphemes(&branch_name, len);
-    // The truncation symbol should only be added if we truncated
-    let truncated_and_symbol = if len < graphemes_len(&branch_name) {
-        let truncation_symbol = get_graphemes(config.truncation_symbol, 1);
-        truncated_graphemes + truncation_symbol.as_str()
+    let branch_graphemes = truncate_text(&branch_name, len, config.truncation_symbol);
+    let topic_graphemes = if let Ok(topic) = get_hg_topic_name(repo_root) {
+        truncate_text(&topic, len, config.truncation_symbol)
     } else {
-        truncated_graphemes
+        String::new()
     };
 
     let parsed = StringFormatter::new(config.format).and_then(|formatter| {
@@ -57,7 +54,8 @@ pub fn module<'a>(context: &'a Context) -> Option<Module<'a>> {
                 _ => None,
             })
             .map(|variable| match variable {
-                "branch" => Some(Ok(truncated_and_symbol.as_str())),
+                "branch" => Some(Ok(branch_graphemes.as_str())),
+                "topic" => Some(Ok(topic_graphemes.as_str())),
                 _ => None,
             })
             .parse(None, Some(context))
@@ -74,26 +72,19 @@ pub fn module<'a>(context: &'a Context) -> Option<Module<'a>> {
     Some(module)
 }
 
-fn get_hg_branch_name(ctx: &Context) -> String {
-    std::fs::read_to_string(ctx.current_dir.join(".hg").join("branch"))
-        .map_or_else(|_| "default".to_string(), |s| s.trim().into())
+fn get_hg_branch_name(hg_root: &Path) -> Result<String, Error> {
+    match read_file(hg_root.join(".hg").join("branch")) {
+        Ok(b) => Ok(b.trim().to_string()),
+        Err(e) => Err(e),
+    }
 }
 
-fn get_hg_current_bookmark(ctx: &Context) -> Option<String> {
-    std::fs::read_to_string(ctx.current_dir.join(".hg").join("bookmarks.current"))
-        .map(|s| s.trim().into())
-        .ok()
+fn get_hg_current_bookmark(hg_root: &Path) -> Result<String, Error> {
+    read_file(hg_root.join(".hg").join("bookmarks.current"))
 }
 
-fn get_graphemes(text: &str, length: usize) -> String {
-    UnicodeSegmentation::graphemes(text, true)
-        .take(length)
-        .collect::<Vec<&str>>()
-        .concat()
-}
-
-fn graphemes_len(text: &str) -> usize {
-    UnicodeSegmentation::graphemes(text, true).count()
+fn get_hg_topic_name(hg_root: &Path) -> Result<String, Error> {
+    read_file(hg_root.join(".hg").join("topic"))
 }
 
 #[cfg(test)]
@@ -183,6 +174,26 @@ mod tests {
             None,
             &[Expect::BranchName("bookmark-101"), Expect::NoTruncation],
         );
+        tempdir.close()
+    }
+
+    #[test]
+    #[ignore]
+    fn test_hg_topic() -> io::Result<()> {
+        let tempdir = fixture_repo(FixtureProvider::Hg)?;
+        let repo_dir = tempdir.path();
+        fs::write(repo_dir.join(".hg").join("topic"), "feature")?;
+
+        let actual = ModuleRenderer::new("hg_branch")
+            .path(repo_dir.to_str().unwrap())
+            .config(toml::toml! {
+                [hg_branch]
+                format = "$topic"
+                disabled = false
+            })
+            .collect();
+
+        assert_eq!(Some(String::from("feature")), actual);
         tempdir.close()
     }
 
@@ -283,7 +294,7 @@ mod tests {
 
     fn expect_hg_branch_with_config(
         repo_dir: &Path,
-        config: Option<toml::Value>,
+        config: Option<toml::Table>,
         expectations: &[Expect],
     ) {
         let actual = ModuleRenderer::new("hg_branch")
@@ -326,8 +337,7 @@ mod tests {
         let expected = Some(format!(
             "on {} ",
             expect_style.paint(format!(
-                "{} {}{}",
-                expect_symbol, expect_branch_name, expect_truncation_symbol
+                "{expect_symbol} {expect_branch_name}{expect_truncation_symbol}"
             )),
         ));
         assert_eq!(expected, actual);
