@@ -77,7 +77,7 @@ fn get_type(os: &os_info::Info) -> Option<String> {
 fn get_version(os: &os_info::Info) -> Option<String> {
     Some(os.version())
         .filter(|&x| x != &os_info::Version::Unknown)
-        .map(|x| x.to_string())
+        .map(os_info::Version::to_string)
 }
 
 #[cfg(test)]
@@ -128,7 +128,7 @@ mod tests {
 
     #[test]
     fn get_symbol_default() {
-        let config = OSConfig::try_load(None);
+        let config = OSConfig::default();
 
         let type_expected_pairs = [
             (Type::Alpine, Some("🏔️ ")),
@@ -189,10 +189,10 @@ mod tests {
             EndeavourOS = " "
             Fedora = " "
             FreeBSD = " "
-            Garuda = "﯑ "
+            Garuda = "󰛓 "
             Gentoo = " "
-            HardenedBSD = "ﲊ "
-            Illumos = " "
+            HardenedBSD = "󰞌 "
+            Illumos = "󰈸 "
             Linux = " "
             Macos = " "
             Manjaro = " "
@@ -201,19 +201,19 @@ mod tests {
             Mint = " "
             NetBSD = " "
             NixOS = " "
-            OpenBSD = " "
+            OpenBSD = "󰈺 "
             SUSE = " "
-            OracleLinux = " "
+            OracleLinux = "󰌷 "
             Pop = " "
             Raspbian = " "
             Redhat = " "
             RedHatEnterprise = " "
-            Redox = " "
-            Solus = "ﴱ "
+            Redox = "󰀘 "
+            Solus = "󰠳 "
             openSUSE = " "
             Ubuntu = " "
             Unknown = " "
-            Windows = " "
+            Windows = "󰍲 "
         };
 
         let config = OSConfig::load(&config_toml);
@@ -230,10 +230,10 @@ mod tests {
             (Type::EndeavourOS, Some(" ")),
             (Type::Fedora, Some(" ")),
             (Type::FreeBSD, Some(" ")),
-            (Type::Garuda, Some("﯑ ")),
+            (Type::Garuda, Some("󰛓 ")),
             (Type::Gentoo, Some(" ")),
-            (Type::HardenedBSD, Some("ﲊ ")),
-            (Type::Illumos, Some(" ")),
+            (Type::HardenedBSD, Some("󰞌 ")),
+            (Type::Illumos, Some("󰈸 ")),
             (Type::Linux, Some(" ")),
             (Type::Macos, Some(" ")),
             (Type::Manjaro, Some(" ")),
@@ -242,19 +242,19 @@ mod tests {
             (Type::Mint, Some(" ")),
             (Type::NetBSD, Some(" ")),
             (Type::NixOS, Some(" ")),
-            (Type::OpenBSD, Some(" ")),
+            (Type::OpenBSD, Some("󰈺 ")),
             (Type::SUSE, Some(" ")),
-            (Type::OracleLinux, Some(" ")),
+            (Type::OracleLinux, Some("󰌷 ")),
             (Type::Pop, Some(" ")),
             (Type::Raspbian, Some(" ")),
             (Type::Redhat, Some(" ")),
             (Type::RedHatEnterprise, Some(" ")),
-            (Type::Redox, Some(" ")),
-            (Type::Solus, Some("ﴱ ")),
+            (Type::Redox, Some("󰀘 ")),
+            (Type::Solus, Some("󰠳 ")),
             (Type::openSUSE, Some(" ")),
             (Type::Ubuntu, Some(" ")),
             (Type::Unknown, Some(" ")),
-            (Type::Windows, Some(" ")),
+            (Type::Windows, Some("󰍲 ")),
         ];
 
         for (t, e) in type_expected_pairs {
@@ -331,10 +331,11 @@ mod tests {
         // - dosc/.vuepress/public/presets/toml/nerd-font-symbols.toml
         // - .github/config-schema.json
         let _ = |t: Type| match t {
+            Type::Alpaquita => "🔔 ",
             Type::Alpine => "🏔️ ",
             Type::Amazon => "🙂 ",
             Type::Android => "🤖 ",
-            Type::Arch => "🎗️ ",
+            Type::Arch | Type::Artix => "🎗️ ",
             Type::CentOS => "💠 ",
             Type::Debian => "🌀 ",
             Type::DragonFly => "🐉 ",
@@ -347,6 +348,7 @@ mod tests {
             Type::HardenedBSD => "🛡️ ",
             Type::Illumos => "🐦 ",
             Type::Linux => "🐧 ",
+            Type::Mabox => "📦 ",
             Type::Macos => "🍎 ",
             Type::Manjaro => "🥭 ",
             Type::Mariner => "🌊 ",
@@ -355,6 +357,8 @@ mod tests {
             Type::NetBSD => "🚩 ",
             Type::NixOS => "❄️ ",
             Type::OpenBSD => "🐡 ",
+            Type::OpenCloudOS => "☁️ ",
+            Type::openEuler => "🦉 ",
             Type::openSUSE => "🦎 ",
             Type::OracleLinux => "🦴 ",
             Type::Pop => "🍭 ",
